all non-driver code now compiles in with g++
