+++ STM32F405RGTx/Core/Inc/bms_entry.hpp
