--- STM32F405RGTx/Core/Inc/bms_tests.h
@@ -1,7 +1,0 @@
-#pragma once
-
-void blinky_loop();
-void delay_us_test();
-void ltc6813_comm_test();
-void ltc6813_adc_test();
-
