+++ STM32F405RGTx/Core/Inc/lut.hpp
@@ -5,4 +5,3 @@
 // lookup tables
 extern uint16_t CRC15_LUT[256];
 extern float ADC_TO_TEMP_LUT[4096];
-
