+++ STM32F405RGTx/Core/Inc/state_machine.hpp
@@ -1,20 +1,19 @@
 /*
- * state_machine.h
+ * state_machine.hpp
  *
  *    Created on: Jul. 11, 2021
  *            Author: tiffanywang
  */
 
-#ifndef _STATE_MACHINE_H_
-#define _STATE_MACHINE_H_
+#pragma once
 
 #include "can.h"
 #include "cmsis_os.h"
 
-#define TURN_ON_PRECHARGE_PIN() (HAL_GPIO_WritePin(PRECHARGE_GPIO_Port, PRECHARGE_Pin, 1))
-#define TURN_OFF_PRECHARGE_PIN() (HAL_GPIO_WritePin(PRECHARGE_GPIO_Port, PRECHARGE_Pin, 0))
-#define TURN_ON_CONTACTOR_PIN() (HAL_GPIO_WritePin(CONTACTOR_GPIO_Port, CONTACTOR_Pin, 1))
-#define TURN_OFF_CONTACTOR_PIN() (HAL_GPIO_WritePin(CONTACTOR_GPIO_Port, CONTACTOR_Pin, 0))
+#define TURN_ON_PRECHARGE_PIN() ( HAL_GPIO_WritePin(PRECHARGE_GPIO_Port, PRECHARGE_Pin, (GPIO_PinState)(1)) )
+#define TURN_OFF_PRECHARGE_PIN() ( HAL_GPIO_WritePin(PRECHARGE_GPIO_Port, PRECHARGE_Pin, (GPIO_PinState)(0)) )
+#define TURN_ON_CONTACTOR_PIN() ( HAL_GPIO_WritePin(CONTACTOR_GPIO_Port, CONTACTOR_Pin, (GPIO_PinState)(1)) )
+#define TURN_OFF_CONTACTOR_PIN() ( HAL_GPIO_WritePin(CONTACTOR_GPIO_Port, CONTACTOR_Pin, (GPIO_PinState)(0)) )
 
 //////////////////////////////////////////////////////////////
 // PACK PARAMETERS
@@ -99,6 +98,3 @@ void StartStateMachine(void *argument);
 
 void StartMeasurements(void *argument);
 
-
-
-#endif /* _STATE_MACHINE_H_ */
